requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2
asyncpg
pydantic==1.10.7
passlib==1.7.4
bcrypt==3.2.0
PyJWT
redis>=4.2.0
APScheduler
python-multipart==0.0.5
greenlet
celery
qrcode
Pillow


